@@ -6,38 +6,42 @@
     "start": "ng serve",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test"
+    "test": "ng test",
+    "lint": "ng lint"
   },
   "private": true,
   "engines": {
     "node": ">=20"
   },
   "dependencies": {
-    "@angular/animations": "20.2.4",
-    "@angular/common": "20.2.4",
-    "@angular/compiler": "20.2.4",
-    "@angular/core": "20.2.4",
-    "@angular/forms": "20.2.4",
-    "@angular/material": "20.2.4",
-    "@angular/cdk": "20.2.4",
-    "@angular/platform-browser": "20.2.4",
-    "@angular/platform-browser-dynamic": "20.2.4",
-    "@angular/router": "20.2.4",
+    "@angular/animations": "21.1.4",
+    "@angular/cdk": "21.1.4",
+    "@angular/common": "21.1.4",
+    "@angular/compiler": "21.1.4",
+    "@angular/core": "21.1.4",
+    "@angular/forms": "21.1.4",
+    "@angular/material": "21.1.4",
+    "@angular/platform-browser": "21.1.4",
+    "@angular/platform-browser-dynamic": "21.1.4",
+    "@angular/router": "21.1.4",
     "rxjs": "~7.8.1",
     "tslib": "^2.7.0",
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "20.2.2",
-    "@angular/cli": "20.2.2",
-    "@angular/compiler-cli": "20.2.4",
+    "@angular-devkit/build-angular": "21.1.4",
+    "@angular/cli": "21.1.4",
+    "@angular/compiler-cli": "21.1.4",
     "@types/jasmine": "~5.1.0",
+    "angular-eslint": "21.2.0",
+    "eslint": "^9.39.2",
     "jasmine-core": "~5.2.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "~5.9.3"
+    "typescript": "~5.9.3",
+    "typescript-eslint": "8.50.1"
   }
 }
