@@ -1,5 +1,9 @@
 from django.test import TestCase, override_settings
 from rest_framework.test import APIClient
+from unittest.mock import patch
+
+from api import rbac
+from api.authentication import Principal
 
 
 class HealthEndpointTests(TestCase):
@@ -39,10 +43,14 @@ class AuthWhoAmITests(TestCase):
         DEBUG=True,
         AUTH_USE_DB_RBAC=False,
     )
-    def test_whoami_forbidden_without_permission(self) -> None:
+    def test_whoami_allows_without_needs_list_permission(self) -> None:
         response = self.client.get("/api/v1/auth/whoami/")
 
-        self.assertEqual(response.status_code, 403)
+        self.assertEqual(response.status_code, 200)
+        body = response.json()
+        self.assertEqual(body["user_id"], "dev-user")
+        self.assertEqual(body["roles"], ["VIEWER"])
+        self.assertEqual(body["permissions"], [])
 
     @override_settings(
         AUTH_ENABLED=False,
@@ -60,3 +68,58 @@ class AuthWhoAmITests(TestCase):
         body = response.json()
         self.assertEqual(body["user_id"], "dev-user")
         self.assertIn("replenishment.needs_list.preview", body["permissions"])
+
+
+class RbacResolutionTests(TestCase):
+    @patch(
+        "api.rbac._fetch_permissions_for_role_codes",
+        return_value={"replenishment.needs_list.approve"},
+    )
+    @patch("api.rbac._resolve_user_id", return_value=None)
+    @patch("api.rbac._db_rbac_enabled", return_value=True)
+    def test_db_rbac_resolves_permissions_from_claim_roles(
+        self,
+        _mock_db_enabled,
+        _mock_user_id,
+        mock_permissions_for_roles,
+    ) -> None:
+        request = type("Request", (), {})()
+        principal = Principal(
+            user_id=None,
+            username="keycloak-user",
+            roles=["ODPEM_DIR_PEOD"],
+            permissions=[],
+        )
+
+        roles, permissions = rbac.resolve_roles_and_permissions(request, principal)
+
+        self.assertIn("ODPEM_DIR_PEOD", roles)
+        self.assertIn("replenishment.needs_list.approve", permissions)
+        self.assertEqual(mock_permissions_for_roles.call_count, 1)
+
+    @patch(
+        "api.rbac._fetch_permissions_for_role_codes",
+        return_value={
+            "replenishment.needs_list.preview",
+            "replenishment.needs_list.create_draft",
+            "replenishment.needs_list.edit_lines",
+        },
+    )
+    @patch("api.rbac._resolve_user_id", return_value=None)
+    @patch("api.rbac._db_rbac_enabled", return_value=True)
+    def test_db_rbac_applies_submit_compat_override_for_logistics_officer(
+        self,
+        _mock_db_enabled,
+        _mock_user_id,
+        _mock_permissions_for_roles,
+    ) -> None:
+        request = type("Request", (), {})()
+        principal = Principal(
+            user_id=None,
+            username="logistics-officer",
+            roles=["TST_LOGISTICS_OFFICER"],
+            permissions=[],
+        )
+
+        _roles, permissions = rbac.resolve_roles_and_permissions(request, principal)
+        self.assertIn("replenishment.needs_list.submit", permissions)
