@@ -4,22 +4,48 @@ from replenishment.views import (
     get_active_event,
     get_all_warehouses,
     needs_list_approve,
+    needs_list_bulk_delete,
+    needs_list_bulk_submit,
     needs_list_cancel,
     needs_list_draft,
     needs_list_edit_lines,
     needs_list_escalate,
+    needs_list_fulfillment_sources,
+    needs_list_generate_transfers,
     needs_list_get,
+    needs_list_list,
     needs_list_mark_completed,
     needs_list_mark_dispatched,
     needs_list_mark_received,
+    needs_list_my_submissions,
     needs_list_preview,
     needs_list_preview_multi,
     needs_list_reject,
+    needs_list_review_reminder,
     needs_list_return,
     needs_list_review_comments,
-    needs_list_review_start,
+    needs_list_summary_version,
     needs_list_start_preparation,
     needs_list_submit,
+    needs_list_transfer_confirm,
+    needs_list_transfer_update,
+    needs_list_transfers,
+    needs_list_donations,
+    needs_list_donations_allocate,
+    needs_list_donations_export,
+    needs_list_procurement_export,
+    # Procurement views
+    procurement_list_create,
+    procurement_detail,
+    procurement_submit,
+    procurement_approve,
+    procurement_reject,
+    procurement_mark_ordered,
+    procurement_mark_shipped,
+    procurement_receive,
+    procurement_cancel,
+    supplier_list_create,
+    supplier_detail,
 )
 
 urlpatterns = [
@@ -28,7 +54,21 @@ urlpatterns = [
     path("needs-list/preview", needs_list_preview, name="needs_list_preview"),
     path("needs-list/preview-multi", needs_list_preview_multi, name="needs_list_preview_multi"),
     path("needs-list/draft", needs_list_draft, name="needs_list_draft"),
+    path("needs-list/my-submissions/", needs_list_my_submissions, name="needs_list_my_submissions"),
+    path("needs-list/bulk-submit/", needs_list_bulk_submit, name="needs_list_bulk_submit"),
+    path("needs-list/bulk-delete/", needs_list_bulk_delete, name="needs_list_bulk_delete"),
+    path("needs-list/", needs_list_list, name="needs_list_list"),
     path("needs-list/<str:needs_list_id>", needs_list_get, name="needs_list_get"),
+    path(
+        "needs-list/<str:needs_list_id>/summary-version",
+        needs_list_summary_version,
+        name="needs_list_summary_version",
+    ),
+    path(
+        "needs-list/<str:needs_list_id>/fulfillment-sources",
+        needs_list_fulfillment_sources,
+        name="needs_list_fulfillment_sources",
+    ),
     path(
         "needs-list/<str:needs_list_id>/lines",
         needs_list_edit_lines,
@@ -43,11 +83,6 @@ urlpatterns = [
         "needs-list/<str:needs_list_id>/submit",
         needs_list_submit,
         name="needs_list_submit",
-    ),
-    path(
-        "needs-list/<str:needs_list_id>/review/start",
-        needs_list_review_start,
-        name="needs_list_review_start",
     ),
     path(
         "needs-list/<str:needs_list_id>/return",
@@ -68,6 +103,11 @@ urlpatterns = [
         "needs-list/<str:needs_list_id>/escalate",
         needs_list_escalate,
         name="needs_list_escalate",
+    ),
+    path(
+        "needs-list/<str:needs_list_id>/review/reminder",
+        needs_list_review_reminder,
+        name="needs_list_review_reminder",
     ),
     path(
         "needs-list/<str:needs_list_id>/start-preparation",
@@ -94,4 +134,57 @@ urlpatterns = [
         needs_list_cancel,
         name="needs_list_cancel",
     ),
+    path(
+        "needs-list/<str:needs_list_id>/generate-transfers",
+        needs_list_generate_transfers,
+        name="needs_list_generate_transfers",
+    ),
+    path(
+        "needs-list/<str:needs_list_id>/transfers",
+        needs_list_transfers,
+        name="needs_list_transfers",
+    ),
+    path(
+        "needs-list/<str:needs_list_id>/transfers/<int:transfer_id>",
+        needs_list_transfer_update,
+        name="needs_list_transfer_update",
+    ),
+    path(
+        "needs-list/<str:needs_list_id>/transfers/<int:transfer_id>/confirm",
+        needs_list_transfer_confirm,
+        name="needs_list_transfer_confirm",
+    ),
+    path(
+        "needs-list/<str:needs_list_id>/donations",
+        needs_list_donations,
+        name="needs_list_donations",
+    ),
+    path(
+        "needs-list/<str:needs_list_id>/donations/allocate",
+        needs_list_donations_allocate,
+        name="needs_list_donations_allocate",
+    ),
+    path(
+        "needs-list/<str:needs_list_id>/donations/export",
+        needs_list_donations_export,
+        name="needs_list_donations_export",
+    ),
+    path(
+        "needs-list/<str:needs_list_id>/procurement/export",
+        needs_list_procurement_export,
+        name="needs_list_procurement_export",
+    ),
+    # ── Procurement (Horizon C) ──────────────────────────────────────────
+    path("procurement/", procurement_list_create, name="procurement_list_create"),
+    path("procurement/<int:procurement_id>", procurement_detail, name="procurement_detail"),
+    path("procurement/<int:procurement_id>/submit", procurement_submit, name="procurement_submit"),
+    path("procurement/<int:procurement_id>/approve", procurement_approve, name="procurement_approve"),
+    path("procurement/<int:procurement_id>/reject", procurement_reject, name="procurement_reject"),
+    path("procurement/<int:procurement_id>/order", procurement_mark_ordered, name="procurement_mark_ordered"),
+    path("procurement/<int:procurement_id>/ship", procurement_mark_shipped, name="procurement_mark_shipped"),
+    path("procurement/<int:procurement_id>/receive", procurement_receive, name="procurement_receive"),
+    path("procurement/<int:procurement_id>/cancel", procurement_cancel, name="procurement_cancel"),
+    # ── Suppliers ────────────────────────────────────────────────────────
+    path("suppliers/", supplier_list_create, name="supplier_list_create"),
+    path("suppliers/<int:supplier_id>", supplier_detail, name="supplier_detail"),
 ]
