@@ -1,5 +1,5 @@
-import { CommonModule } from '@angular/common';
-import { Component, Inject } from '@angular/core';
+
+import { Component, inject } from '@angular/core';
 import { FormControl, ReactiveFormsModule, Validators } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialogModule, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
@@ -15,24 +15,23 @@ export interface PhaseSelectDialogData {
   selector: 'app-phase-select-dialog',
   standalone: true,
   imports: [
-    CommonModule,
     ReactiveFormsModule,
     MatDialogModule,
     MatFormFieldModule,
     MatSelectModule,
     MatButtonModule
-  ],
+],
   templateUrl: './phase-select-dialog.component.html',
   styleUrl: './phase-select-dialog.component.scss'
 })
 export class PhaseSelectDialogComponent {
+  private dialogRef = inject<MatDialogRef<PhaseSelectDialogComponent, EventPhase>>(MatDialogRef);
+  data = inject<PhaseSelectDialogData>(MAT_DIALOG_DATA);
+
   readonly phaseOptions: EventPhase[] = ['SURGE', 'STABILIZED', 'BASELINE'];
   phaseControl: FormControl<EventPhase | null>;
 
-  constructor(
-    private dialogRef: MatDialogRef<PhaseSelectDialogComponent, EventPhase>,
-    @Inject(MAT_DIALOG_DATA) public data: PhaseSelectDialogData
-  ) {
+  constructor() {
     this.phaseControl = new FormControl<EventPhase | null>(this.data.currentPhase ?? null, {
       nonNullable: false,
       validators: [Validators.required]
